requirements.txt
# Core
Django==5.1.3
django-environ==0.11.2
gunicorn==23.0.0
psycopg2==3.2.3

# Rest apis
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.6.0
drf-yasg==1.21.8
whitenoise==6.8.2

# Certificates
cryptography==43.0.3
